# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "api 计算Bestdori谱面难度，获得Bestdori数据，常见Bandori谱面格式转换等",
        "title": "ayachan API",
        "contact": {
            "name": "6QHTSK"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://mit-license.org/"
        },
        "version": "2.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/v2",
    "paths": {
...
        "/map-data": {
            "post": {
                "description": "根据用户上传的谱面获取指定格式谱面",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mapData"
                ],
                "summary": "从用户上传谱面获取谱面",
                "parameters": [
                    {
                        "description": "需转换的谱面",
                        "name": "map",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.MapInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/utils.MapOutput"
                        }
                    }
                }
            }
        },
        "/map-data/bestdori/:chartID": {
            "get": {
                "description": "根据谱面ID、谱面难度从Bestdori获取谱面",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mapData"
                ],
                "summary": "从Bestdori获取谱面",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "谱面ID",
                        "name": "chartID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "谱面难度差分指定，默认Expert",
                        "name": "diff",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "输出谱面格式，默认BestdoriV2",
                        "name": "map_format_out",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/utils.MapOutput"
                        }
                    }
                }
            }
        },
        "/map-info": {
            "post": {
                "description": "根据用户上传的谱面获取指定格式谱面信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mapInfo"
                ],
                "summary": "从用户上传谱面获取谱面信息",
                "parameters": [
                    {
                        "description": "需转换的谱面",
                        "name": "map",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.MapInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "对应的谱面",
                        "schema": {
                            "$ref": "#/definitions/Controllers.MapInfoOutput"
                        }
                    }
                }
            }
        },
        "/map-info/bestdori/:chartID": {
            "get": {
                "description": "根据谱面ID、谱面难度从Bestdori获取谱面并分析得到数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mapInfo"
                ],
                "summary": "从Bestdori获取谱面并分析数据",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "谱面ID",
                        "name": "chartID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "谱面难度差分指定，默认Expert",
                        "name": "diff",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "输出谱面格式，默认BestdoriV2",
                        "name": "map_format_out",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "解析的谱面信息",
                        "schema": {
                            "$ref": "#/definitions/Controllers.MapInfoOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Controllers.MapInfoOutput": {
            "type": "object",
            "properties": {
                "mapDifficultyExtend": {
                    "description": "MapDifficultyExtend"
                },
                "mapDifficultyStandard": {
                    "$ref": "#/definitions/Models.MapDifficultyStandard"
                },
                "mapInfoExtend": {
                    "description": "MapInfoExtend"
                },
                "mapInfoStandard": {
                    "$ref": "#/definitions/Models.MapInfoStandard"
                },
                "result": {
                    "type": "boolean"
                }
            }
        },
...
        "Models.Distribution": {
            "type": "object",
            "properties": {
                "hit": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "note": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "Models.MapDifficultyStandard": {
            "type": "object",
            "properties": {
                "blueWhiteDifficulty": {
                    "type": "number"
                },
                "difficulty": {
                    "type": "number"
                },
                "maxScreenNPS": {
                    "type": "number"
                },
                "totalHPS": {
                    "type": "number"
                },
                "totalNPS": {
                    "type": "number"
                }
            }
        },
        "Models.MapInfoStandard": {
            "type": "object",
            "properties": {
                "bpmhigh": {
                    "type": "number"
                },
                "bpmlow": {
                    "type": "number"
                },
                "distribution": {
                    "$ref": "#/definitions/Models.Distribution"
                },
                "irregular": {
                    "description": "存在多压/交叉（出张）0 失败 1 标准 2 非标准",
                    "type": "integer"
                },
                "irregularInfo": {
                    "description": "无法分析的第一个错误情况",
                    "type": "string"
                },
                "mainBPM": {
                    "type": "number"
                },
                "maxScreenNPS": {
                    "type": "number"
                },
                "noteCount": {
                    "$ref": "#/definitions/Models.NoteCount"
                },
                "totalHPS": {
                    "type": "number"
                },
                "totalHitNote": {
                    "type": "integer"
                },
                "totalNPS": {
                    "type": "number"
                },
                "totalNote": {
                    "type": "integer"
                },
                "totalTime": {
                    "type": "number"
                }
            }
        },
        "Models.NoteCount": {
            "type": "object",
            "properties": {
                "direction": {
                    "type": "object",
                    "properties": {
                        "left": {
                            "type": "integer"
                        },
                        "right": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "flick": {
                    "type": "integer"
                },
                "single": {
                    "type": "integer"
                },
                "slideEnd": {
                    "type": "integer"
                },
                "slideHidden": {
                    "type": "integer"
                },
                "slideStart": {
                    "type": "integer"
                },
                "slideTick": {
                    "type": "integer"
                },
                "sprhythm": {
                    "type": "boolean"
                }
            }
        },
...
        "utils.MapInput": {
            "type": "object",
            "properties": {
                "Options": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "map": {},
                "map_format_in": {
                    "type": "string"
                }
            }
        },
        "utils.MapOutput": {
            "type": "object",
            "properties": {
                "map": {},
                "result": {
                    "type": "boolean"
                }
            }
        }
    }
}